+++ NEWS.rst
@@ -7,6 +7,8 @@ of every change, see the Git log.
 Latest
 ------
 * tbd
+* Minor: Added mkspecs for Android x86_64: cxx_android5_clang38_x64 and
+  cxx_android5_clang50_x64.
 
 4.8.0
 -----